+++ plugins/DataRecorder.hpp
@@ -1,21 +1,21 @@
 /**
- * @file SNBWriter.hpp Module to write snb data
+ * @file DataRecorder.hpp Module to record data
  *
  * This is part of the DUNE DAQ , copyright 2020.
  * Licensing/copyright details are in the COPYING file that you should have
  * received with this code.
 */
 
-#ifndef READOUT_PLUGINS_BUFFERED_FILE_STREAMER_HPP_
-#define READOUT_PLUGINS_BUFFERED_FILE_STREAMER_HPP_
+#ifndef READOUT_PLUGINS_DATA_RECORDER_HPP_
+#define READOUT_PLUGINS_DATA_RECORDER_HPP_
 
 #include "appfwk/DAQModule.hpp"
 #include "appfwk/DAQSource.hpp"
 #include "appfwk/ThreadHelper.hpp"
-#include "readout/bufferedfilestreamer/Structs.hpp"
+#include "readout/datarecorder/Structs.hpp"
 #include "readout/ReadoutTypes.hpp"
 #include "ReadoutStatistics.hpp"
-#include "BufferedWriter.hpp"
+#include "BufferedFileWriter.hpp"
 
 #include <iostream>
 #include <fstream>
@@ -24,14 +24,14 @@
 namespace dunedaq {
   namespace readout {
 
-    class BufferedFileStreamer : public dunedaq::appfwk::DAQModule {
+    class DataRecorder : public dunedaq::appfwk::DAQModule {
     public:
-      explicit BufferedFileStreamer(const std::string& name);
+      explicit DataRecorder(const std::string& name);
 
-      BufferedFileStreamer(const BufferedFileStreamer&) = delete;
-      BufferedFileStreamer& operator=(const BufferedFileStreamer&) = delete;
-      BufferedFileStreamer(BufferedFileStreamer&&) = delete;
-      BufferedFileStreamer& operator=(BufferedFileStreamer&&) = delete;
+      DataRecorder(const DataRecorder&) = delete;
+      DataRecorder& operator=(const DataRecorder&) = delete;
+      DataRecorder(DataRecorder&&) = delete;
+      DataRecorder& operator=(DataRecorder&&) = delete;
 
       void init(const nlohmann::json& obj) override;
       void get_info(opmonlib::InfoCollector& ci, int level) override;
@@ -50,8 +50,8 @@ namespace dunedaq {
       int fd;
       std::string m_output_file;
       std::mutex m_start_lock;
-      bufferedfilestreamer::Conf m_conf;
-      BufferedWriter<types::WIB_SUPERCHUNK_STRUCT> m_buffered_writer;
+      datarecorder::Conf m_conf;
+      BufferedFileWriter<types::WIB_SUPERCHUNK_STRUCT> m_buffered_writer;
 
       // Stats
       stats::counter_t m_packets_processed_total{0};
@@ -61,4 +61,4 @@ namespace dunedaq {
   }
 }
 
-#endif //READOUT_PLUGINS_BUFFERED_FILE_STREAMER_HPP_
+#endif //READOUT_PLUGINS_DATA_RECORDER_HPP_
