+++ schema/readout/datarecorderinfo.jsonnet
@@ -1,9 +1,9 @@
-// This is the application info schema used by the buffered file streamer module.
+// This is the application info schema used by the DataRecorder module.
 // It describes the information object structure passed by the application
 // for operational monitoring
 
 local moo = import "moo.jsonnet";
-local s = moo.oschema.schema("dunedaq.readout.bufferedfilestreamerinfo");
+local s = moo.oschema.schema("dunedaq.readout.datarecorderinfo");
 
 local info = {
    cl : s.string("class_s", moo.re.ident,
@@ -14,7 +14,7 @@ local info = {
                   doc="A float of 8 bytes"),
 
    info: s.record("Info", [
-       s.field("class_name", self.cl, "bufferedfilestreamerinfo", doc="Info class name"),
+       s.field("class_name", self.cl, "datarecorderinfo", doc="Info class name"),
        s.field("packets_processed", self.uint8, 0, doc="Number of packets processed"),
        s.field("throughput_processed_packets", self.float8, 0, doc="Throughput of processed packets"),
    ], doc="Data link handler information information")
