+++ src/BufferedFileReader.hpp
@@ -1,12 +1,12 @@
 /**
-* @file BufferedWriter.hpp Writes data to a file
+* @file BufferedFileReader.hpp Reads data to a file
 *
 * This is part of the DUNE DAQ , copyright 2021.
 * Licensing/copyright details are in the COPYING file that you should have
 * received with this code.
 */
-#ifndef READOUT_SRC_BUFFEREDWRITER_HPP_
-#define READOUT_SRC_BUFFEREDWRITER_HPP_
+#ifndef READOUT_SRC_BUFFEREDFILEREADER_HPP_
+#define READOUT_SRC_BUFFEREDFILEREADER_HPP_
 
 #include "ReadoutIssues.hpp"
 #include "readout/ReadoutLogging.hpp"
@@ -33,63 +33,57 @@ using namespace dunedaq::readout::logging;
 
 namespace dunedaq {
   namespace readout {
-    template<class RawType, size_t Alignment = 4096>
-    class BufferedWriter {
-      using io_sink_t = boost::iostreams::file_descriptor_sink;
-      using aligned_allocator_t = boost::alignment::aligned_allocator<io_sink_t::char_type, Alignment>;
-      using filtering_ostream_t = boost::iostreams::filtering_stream<boost::iostreams::output, char, std::char_traits<char>, aligned_allocator_t>;
+    template<class RawType, size_t Alignment=4096>
+    class BufferedFileReader {
+      using io_source_t = boost::iostreams::file_descriptor_source;
+      using aligned_allocator_t = boost::alignment::aligned_allocator<io_source_t::char_type, Alignment>;
+      using filtering_istream_t = boost::iostreams::filtering_stream<boost::iostreams::input, char, std::char_traits<char>, aligned_allocator_t>;
     public:
-      explicit BufferedWriter(std::string filename, size_t buffer_size, std::string compression_algorithm = "None")
+      explicit BufferedFileReader(std::string filename, size_t buffer_size, std::string compression_algorithm = "None")
       {
         open(filename, buffer_size, compression_algorithm);
       }
 
-      explicit BufferedWriter() {
+      explicit BufferedFileReader() {
 
       }
 
-      ~BufferedWriter() {
-        if (m_is_open) close();
-      }
-
-      BufferedWriter(const BufferedWriter&)
-      = delete; ///< BufferedWriter is not copy-constructible
-      BufferedWriter& operator=(const BufferedWriter&)
-      = delete; ///< BufferedWriter is not copy-assginable
-      BufferedWriter(BufferedWriter&&)
-      = delete; ///< BufferedWriter is not move-constructible
-      BufferedWriter& operator=(BufferedWriter&&)
-      = delete; ///< BufferedWriter is not move-assignable
+      BufferedFileReader(const BufferedFileReader&)
+      = delete; ///< BufferedFileReader is not copy-constructible
+      BufferedFileReader& operator=(const BufferedFileReader&)
+      = delete; ///< BufferedFileReader is not copy-assginable
+      BufferedFileReader(BufferedFileReader&&)
+      = delete; ///< BufferedFileReader is not move-constructible
+      BufferedFileReader& operator=(BufferedFileReader&&)
+      = delete; ///< BufferedFileReader is not move-assignable
 
       void open(std::string filename, size_t buffer_size, std::string compression_algorithm = "None") {
-        if (m_is_open) flush();
-
         m_filename = filename;
         m_buffer_size = buffer_size;
         m_compression_algorithm = compression_algorithm;
 
-        m_fd = ::open(m_filename.c_str(), O_CREAT | O_WRONLY | O_DIRECT , 0644);
-        if (m_fd == -1) {
+        int fd = ::open(m_filename.c_str(), O_RDONLY);
+        if (fd == -1) {
           throw CannotOpenFile(ERS_HERE, m_filename);
         }
 
-        m_sink = io_sink_t(m_fd, boost::iostreams::file_descriptor_flags::close_handle);
+        io_source_t io_source(fd, boost::iostreams::file_descriptor_flags::close_handle);
         if (m_compression_algorithm == "zstd") {
           TLOG_DEBUG(TLVL_WORK_STEPS) << "Using zstd compression" << std::endl;
-          m_output_stream.push(boost::iostreams::zstd_compressor(boost::iostreams::zstd::best_speed));
+          m_input_stream.push(boost::iostreams::zstd_decompressor());
         } else if (m_compression_algorithm == "lzma") {
           TLOG_DEBUG(TLVL_WORK_STEPS) << "Using lzma compression" << std::endl;
-          m_output_stream.push(boost::iostreams::lzma_compressor(boost::iostreams::lzma::best_speed));
+          m_input_stream.push(boost::iostreams::lzma_decompressor());
         } else if (m_compression_algorithm == "zlib") {
           TLOG_DEBUG(TLVL_WORK_STEPS) << "Using zlib compression" << std::endl;
-          m_output_stream.push(boost::iostreams::zlib_compressor(boost::iostreams::zlib::best_speed));
+          m_input_stream.push(boost::iostreams::zlib_decompressor());
         } else if (m_compression_algorithm == "None") {
           TLOG_DEBUG(TLVL_WORK_STEPS) << "Running without compression" << std::endl;
         } else {
           throw ConfigurationError(ERS_HERE, "Non-recognized compression algorithm: " + m_compression_algorithm);
         }
 
-        m_output_stream.push(m_sink, m_buffer_size);
+        m_input_stream.push(io_source, m_buffer_size);
         m_is_open = true;
       }
 
@@ -97,37 +91,27 @@ namespace dunedaq {
         return m_is_open;
       }
 
-      bool write(const RawType& element) {
+      bool read(RawType& element) {
         if (!m_is_open) return false;
-        m_output_stream.write((char*)&element, sizeof(element));
-        return !m_output_stream.bad();
+        m_input_stream.read((char*)&element, sizeof(element));
+        return (m_input_stream.gcount() == sizeof(element));
       }
 
       void close() {
-        fcntl(m_fd, F_SETFL, O_CREAT | O_WRONLY);
-        m_output_stream.reset();
+        m_input_stream.reset();
         m_is_open = false;
       }
-
-      void flush() {
-        fcntl(m_fd, F_SETFL, O_CREAT | O_WRONLY);
-        m_output_stream.flush();
-        fcntl(m_fd, F_SETFL, O_CREAT | O_WRONLY | O_DIRECT);
-      }
-
 
     private:
       std::string m_filename;
       size_t m_buffer_size;
       std::string m_compression_algorithm;
 
-      int m_fd;
-      io_sink_t m_sink;
-      filtering_ostream_t m_output_stream;
+      filtering_istream_t m_input_stream;
       bool m_is_open = false;
     };
 
   } // namespace readout
 } // namespace dunedaq
 
-#endif // READOUT_SRC_BUFFEREDWRITER_HPP_
+#endif // READOUT_SRC_BUFFEREDFILEREADER_HPP_
